Move cleanse to helper as it is not actually doing validation.
