+++ src/components/Inputs/BirthdateInput/BirthdateInput.js
@@ -8,7 +8,7 @@ import dayjs from '../../../helpers/getDayjs.js'
 import useErrorMessage from '../../../hooks/useErrorMessage.js'
 import useInputValidation from '../../../hooks/useInputValidation.js'
 import * as Validators from '../../../validators/BirthdateInputValidator'
-import { cleanse } from '../../../validators/NumberValidator.js'
+import cleanse from '../../../helpers/cleanse.js'
 
 const { DATE_FORMATS, dateMaskByFormat, dateStringMatchesFormat } = Validators
