+++ src/helpers/cleanse.js
@@ -1,0 +1,2 @@
+// first replace removes text-mask's underscores, second fixes like: '2//'
+export default (value) => value.replace(/[,_$]/g, '').replace('//', '/')
